feat(Update): patch on Module Master

+++ app/Http/Controllers/Admin/ModulesController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Models\Module;
+use App\Models\ProjectDetail;
 use App\Http\Requests\ModuleRequest;
 
 class ModulesController extends Controller
@@ -33,8 +34,14 @@ class ModulesController extends Controller
     }
 
     public function destroy(Module $module) {
-        $module->delete();
+        $check = ProjectDetail::where('moduleable_type', 'App\Models\Module')->where('moduleable_id', $module->id)->get();
+        
+        if(!$check){
+            $module->delete();
 
-        return redirect()->route('admin.modules.index')->with('success', 'Module deleted successfully');
+            return redirect()->route('admin.modules.index')->with('success', 'Module deleted successfully');
+        }
+        
+        return redirect()->route('admin.modules.index')->with('error', 'This module is still used on a project!');
     }
 }
